backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-bcrypt
